+++ cloud-groep22_functions/src/main/java/cloudnative/fitapp/functions/SetFunctions.java
@@ -4,6 +4,10 @@ import com.microsoft.azure.functions.*;
 import com.microsoft.azure.functions.annotation.*;
 import cloudnative.fitapp.domain.Set;
 import cloudnative.fitapp.service.SetService;
+import cloudnative.fitapp.service.ExerciseService;
+import cloudnative.fitapp.service.WorkoutService;
+import cloudnative.fitapp.service.UserService;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import java.util.List;
 import java.util.Optional;
 
@@ -19,7 +23,7 @@ public class SetFunctions extends BaseFunctionHandler {
     public HttpResponseMessage getAllSets(
             @HttpTrigger(
                     name = "req",
-                    methods = {HttpMethod.GET},
+                    methods = {HttpMethod.GET, HttpMethod.OPTIONS},
                     route = "sets",
                     authLevel = AuthorizationLevel.ANONYMOUS)
             HttpRequestMessage<Optional<String>> request,
@@ -27,9 +31,20 @@ public class SetFunctions extends BaseFunctionHandler {
 
         context.getLogger().info("Getting all sets");
 
+        if (request.getHttpMethod() == HttpMethod.OPTIONS) {
+            return handleCors(request);
+        }
+
         try {
             validateToken(request);
-            SetService setService = getBean(SetService.class);
+
+            // Create services
+            BCryptPasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
+            UserService userService = new UserService(cosmosDBService, passwordEncoder);
+            WorkoutService workoutService = new WorkoutService(cosmosDBService, userService);
+            ExerciseService exerciseService = new ExerciseService(cosmosDBService, workoutService);
+            SetService setService = new SetService(cosmosDBService, exerciseService);
+
             List<Set> sets = setService.getAllSets();
             return createResponse(request, sets);
         } catch (Exception e) {
@@ -44,7 +59,7 @@ public class SetFunctions extends BaseFunctionHandler {
     public HttpResponseMessage getSetById(
             @HttpTrigger(
                     name = "req",
-                    methods = {HttpMethod.GET},
+                    methods = {HttpMethod.GET, HttpMethod.OPTIONS},
                     route = "sets/{id}",
                     authLevel = AuthorizationLevel.ANONYMOUS)
             HttpRequestMessage<Optional<String>> request,
@@ -53,9 +68,20 @@ public class SetFunctions extends BaseFunctionHandler {
 
         context.getLogger().info("Getting set by ID: " + id);
 
+        if (request.getHttpMethod() == HttpMethod.OPTIONS) {
+            return handleCors(request);
+        }
+
         try {
             validateToken(request);
-            SetService setService = getBean(SetService.class);
+
+            // Create services
+            BCryptPasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
+            UserService userService = new UserService(cosmosDBService, passwordEncoder);
+            WorkoutService workoutService = new WorkoutService(cosmosDBService, userService);
+            ExerciseService exerciseService = new ExerciseService(cosmosDBService, workoutService);
+            SetService setService = new SetService(cosmosDBService, exerciseService);
+
             Set set = setService.getSetById(Long.parseLong(id));
 
             if (set == null) {
@@ -75,7 +101,7 @@ public class SetFunctions extends BaseFunctionHandler {
     public HttpResponseMessage addSetToExercise(
             @HttpTrigger(
                     name = "req",
-                    methods = {HttpMethod.POST},
+                    methods = {HttpMethod.POST, HttpMethod.OPTIONS},
                     route = "sets/exercise/{exerciseId}/addSet",
                     authLevel = AuthorizationLevel.ANONYMOUS)
             HttpRequestMessage<String> request,
@@ -84,10 +110,21 @@ public class SetFunctions extends BaseFunctionHandler {
 
         context.getLogger().info("Adding set to exercise: " + exerciseId);
 
+        if (request.getHttpMethod() == HttpMethod.OPTIONS) {
+            return handleCors(request);
+        }
+
         try {
             validateToken(request);
             Set set = parseBody(request, Set.class);
-            SetService setService = getBean(SetService.class);
+
+            // Create services
+            BCryptPasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
+            UserService userService = new UserService(cosmosDBService, passwordEncoder);
+            WorkoutService workoutService = new WorkoutService(cosmosDBService, userService);
+            ExerciseService exerciseService = new ExerciseService(cosmosDBService, workoutService);
+            SetService setService = new SetService(cosmosDBService, exerciseService);
+
             Set newSet = setService.addSetToExercise(Long.parseLong(exerciseId), set);
             return createResponse(request, newSet);
         } catch (Exception e) {
@@ -102,7 +139,7 @@ public class SetFunctions extends BaseFunctionHandler {
     public HttpResponseMessage updateSet(
             @HttpTrigger(
                     name = "req",
-                    methods = {HttpMethod.PUT},
+                    methods = {HttpMethod.PUT, HttpMethod.OPTIONS},
                     route = "sets/{id}",
                     authLevel = AuthorizationLevel.ANONYMOUS)
             HttpRequestMessage<String> request,
@@ -111,10 +148,21 @@ public class SetFunctions extends BaseFunctionHandler {
 
         context.getLogger().info("Updating set: " + id);
 
+        if (request.getHttpMethod() == HttpMethod.OPTIONS) {
+            return handleCors(request);
+        }
+
         try {
             validateToken(request);
             Set set = parseBody(request, Set.class);
-            SetService setService = getBean(SetService.class);
+
+            // Create services
+            BCryptPasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
+            UserService userService = new UserService(cosmosDBService, passwordEncoder);
+            WorkoutService workoutService = new WorkoutService(cosmosDBService, userService);
+            ExerciseService exerciseService = new ExerciseService(cosmosDBService, workoutService);
+            SetService setService = new SetService(cosmosDBService, exerciseService);
+
             Set updatedSet = setService.updateSet(Long.parseLong(id), set);
             return createResponse(request, updatedSet);
         } catch (Exception e) {
@@ -129,7 +177,7 @@ public class SetFunctions extends BaseFunctionHandler {
     public HttpResponseMessage deleteSet(
             @HttpTrigger(
                     name = "req",
-                    methods = {HttpMethod.DELETE},
+                    methods = {HttpMethod.DELETE, HttpMethod.OPTIONS},
                     route = "sets/{id}",
                     authLevel = AuthorizationLevel.ANONYMOUS)
             HttpRequestMessage<Optional<String>> request,
@@ -138,16 +186,28 @@ public class SetFunctions extends BaseFunctionHandler {
 
         context.getLogger().info("Deleting set: " + id);
 
+        if (request.getHttpMethod() == HttpMethod.OPTIONS) {
+            return handleCors(request);
+        }
+
         try {
             validateToken(request);
-            SetService setService = getBean(SetService.class);
+
+            // Create services
+            BCryptPasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
+            UserService userService = new UserService(cosmosDBService, passwordEncoder);
+            WorkoutService workoutService = new WorkoutService(cosmosDBService, userService);
+            ExerciseService exerciseService = new ExerciseService(cosmosDBService, workoutService);
+            SetService setService = new SetService(cosmosDBService, exerciseService);
 
             if (setService.getSetById(Long.parseLong(id)) == null) {
                 return createErrorResponse(request, HttpStatus.NOT_FOUND, "Set not found");
             }
 
             setService.deleteSet(Long.parseLong(id));
-            return request.createResponseBuilder(HttpStatus.NO_CONTENT).build();
+            return request.createResponseBuilder(HttpStatus.NO_CONTENT)
+                    .header("Access-Control-Allow-Origin", "*")
+                    .build();
         } catch (Exception e) {
             return handleException(request, e);
         }
